requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1

qdrant-client==1.9.1

requests==2.31.0
pydantic==2.7.1
pydantic-core==2.18.2

tqdm==4.66.4
python-dotenv==1.0.1
requests==2.31.0
pytest==7.3.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-mock==3.10.0